유저 정보로 가져오기 (get, set)

+++ Assets/Script/UI/UIManager.cs
@@ -23,7 +23,7 @@ public class UIManager : MonoBehaviour
         }
     }
 
-    public void SetMoneyText(int value)
+    public void SetMoneyText(long value)
     {
         moneyText.text = value.ToString();
     }
